refactor: move org sidebar to sidebar folder
