+++ wondrous-app/components/Common/Sidebar/Organization/AboutEntity/PrivacyIcon.tsx
