+++ wondrous-app/components/ListViewAdmin/ColumnEntry.tsx
@@ -24,7 +24,7 @@ import { SnackbarAlertContext } from 'components/Common/SnackbarAlert';
 import { LinkIcon, DueDateIcon } from 'components/Icons/taskModalIcons';
 import { Rewards } from 'components/Common/TaskViewModal/helpers';
 import { ListViewItemBodyWrapper, ListViewItemDataContainer, ListViewItemActions } from 'components/ListView/styles';
-import { NoLogoDAO } from 'components/Common/Sidebar/SidebarPermanent/styles';
+import { NoLogoDAO } from 'components/Common/Sidebar/Permanent/styles';
 import { SafeImage } from 'components/Common/Image';
 import DefaultUserImage from 'components/Common/Image/DefaultUserImage';
 import { DAOIcon } from 'components/Icons/dao';
